refactor(test-support): move secrets test fixtures behind feature
Move shared test credential and certificate helpers into the secrets crate,
expose them via a test-support feature, and update dependent test fixtures to
import them from the new location. Also extract the api-core default test config
into test_support for reuse.
