+++ crates/api-core/src/tests/power_shelf_state_controller/maintenance.rs
@@ -44,7 +44,8 @@ use carbide_power_shelf_controller::metrics::PowerShelfMetrics;
 use carbide_uuid::power_shelf::PowerShelfId;
 use carbide_uuid::rack::RackId;
 use db::{expected_power_shelf as db_expected_power_shelf, power_shelf as db_power_shelf};
-use forge_secrets::credentials::{Credentials, TestCredentialManager};
+use forge_secrets::credentials::Credentials;
+use forge_secrets::test_support::credentials::TestCredentialManager;
 use librms::protos::rack_manager as rms;
 use mac_address::MacAddress;
 use model::expected_power_shelf::ExpectedPowerShelf;
