+++ crates/api-core/src/tests/power_shelf_state_controller/mod.rs
@@ -34,7 +34,7 @@ mod error_state;
 mod fixtures;
 mod maintenance;
 use fixtures::power_shelf::{mark_power_shelf_as_deleted, set_power_shelf_controller_state};
-use forge_secrets::credentials::TestCredentialManager;
+use forge_secrets::test_support::credentials::TestCredentialManager;
 
 #[crate::sqlx_test]
 async fn test_power_shelf_state_transition_validation(
