+++ crates/bmc-explorer-cli/Cargo.toml
@@ -6,15 +6,11 @@ license.workspace = true
 authors.workspace = true
 
 [dependencies]
-# [local-dependencies]
-# DO NOT PUT DEPENDENCIES OTHER THAN LOCAL DEPS HERE, THEY SHOULD ALL HAVE 'path =' IN THEM.
-carbide-site-explorer = { path = "../site-explorer" }
-carbide-secrets = { path = "../secrets" }
 carbide-ipmi = { path = "../ipmi" }
 carbide-redfish = { path = "../redfish" }
-# DO NOT PUT DEPENDENCIES OTHER THAN LOCAL DEPS HERE, THEY SHOULD ALL HAVE 'path =' IN THEM
+carbide-secrets = { path = "../secrets", features = ["test-support"] }
+carbide-site-explorer = { path = "../site-explorer" }
 
-#these are alphabetized
 arc-swap = { workspace = true }
 clap = { workspace = true }
 libredfish = { workspace = true }
