+++ crates/bmc-explorer-cli/src/main.rs
@@ -24,7 +24,8 @@ use carbide_redfish::nv_redfish::NvRedfishClientPool;
 use carbide_site_explorer::BmcEndpointExplorer;
 use carbide_site_explorer::config::SiteExplorerExploreMode;
 use clap::Parser;
-use forge_secrets::credentials::{Credentials, TestCredentialManager};
+use forge_secrets::credentials::Credentials;
+use forge_secrets::test_support::credentials::TestCredentialManager;
 use mac_address::MacAddress;
 use tracing_subscriber::fmt;
