+++ crates/redfish/src/libredfish/test_support.rs
@@ -23,7 +23,8 @@ use std::time::Duration;
 
 use async_trait::async_trait;
 use chrono::Utc;
-use forge_secrets::credentials::{CredentialReader, TestCredentialManager};
+use forge_secrets::credentials::CredentialReader;
+use forge_secrets::test_support::credentials::TestCredentialManager;
 use libredfish::model::certificate::Certificate;
 use libredfish::model::component_integrity::{ComponentIntegrities, ComponentIntegrity};
 use libredfish::model::oem::nvidia_dpu::{HostPrivilegeLevel, NicMode};
