+++ crates/secrets/Cargo.toml
@@ -25,6 +25,10 @@ authors.workspace = true
 [lib]
 name = "forge_secrets"
 
+[features]
+default = []
+test-support = []
+
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
